requirements.txt
beautifulsoup4==4.7.1
Bootstrap-Flask==1.0.10
bs4==0.0.1
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
codecov==2.0.15
Flask==1.0.2
GitPython==2.1.11
lxml==4.3.2
pytest==4.3.0
pytest-cov==2.6.1
python-dotenv==0.10.1
pytz==2018.9
ruamel.yaml==0.15.89

